Use centos6 instead of centos5
After moving jira/confluence/crowd to our AWS installation, we started
using fastly. But fastly does not support SSLv3 for HTTPS traffic.

This is a problem for the current container used to authenticate jenkins
using the crowd apache plugin. That container is built using centos5,
which by default uses SSLv3.

The container code is
here:https://github.com/mergermarket/apache_with_mod_authnz_crowd built
automatically by docker here:
https://hub.docker.com/r/mergermarket/apache-with-mod-authnz-crowd/ and
configured in the jenkins stack:
https://github.com/mergermarket/jenkins/blob/master/infrastructure/ecs-deployment-scripts/default-taskdef.jq#L52

The apache configuration is a simple file dropped in the jenkins volume,
and it is not in source control.

We want to rebuild that container with a version that support something
different than SSLv3. The problem is that centos5 is not supported
anymore.

We found a binary for redhat6 and we will try with that binary[1]

[1] https://downloads.atlassian.com/software/crowd/downloads/cwdapache/packages/rhel6/mod_authnz_crowd-2.2.2-1.el6.x86_64.rpm